convert component
